Move GrpcServlet to autoconfiguration
